+++ spring-addons-starter-oidc/src/main/java/com/c4_soft/springaddons/security/oidc/starter/properties/SpringAddonsOidcClientProperties.java
@@ -8,14 +8,7 @@ import java.util.Optional;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager;
-import org.springframework.web.client.RestClient;
-import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import com.c4_soft.springaddons.security.oidc.starter.rest.BearerProvider;
-import com.c4_soft.springaddons.security.oidc.starter.rest.DefaultBearerProvider;
 
 import lombok.Data;
 
@@ -178,11 +171,6 @@ public class SpringAddonsOidcClientProperties {
     private Map<String, List<RequestParam>> tokenRequestParams = new HashMap<>();
 
     /**
-     * Configuration for the SpringAddonsRestSupport
-     */
-    private Map<String, RestProperties> rest = new HashMap<>();
-
-    /**
      * Logout properties for OpenID Providers which do not implement the RP-Initiated Logout spec
      *
      * @author Jerome Wacongne ch4mp&#64;c4-soft.com
@@ -262,40 +250,6 @@ public class SpringAddonsOidcClientProperties {
          * Status for the response after BFF logout, with location to authorization server logout endpoint
          */
         private HttpStatus rpInitiatedLogout = HttpStatus.FOUND;
-    }
-
-    @Data
-    @ConfigurationProperties
-    public static class RestProperties {
-        /**
-         * Base URI used to build the REST client ({@link RestClient} or {@link WebClient})
-         */
-        private URI baseUri;
-
-        /**
-         * <p>
-         * If provided, it is used to get an access token from the {@link OAuth2AuthorizedClientManager}.
-         * </p>
-         * <p>
-         * Must reference a valid entry under spring.security.oauth2.client.registration
-         * </p>
-         * <p>
-         * Mutually exclusive with forward-bearer property.
-         * </p>
-         */
-        private Optional<String> auth2RegistrationId = Optional.empty();
-
-        /**
-         * <p>
-         * If true, a {@link BearerProvider} is used to retrieve a Bearer token from the {@link Authentication} in the security context.
-         * </p>
-         * <p>
-         * Mutually exclusive with auth2-registration-id property.
-         * </p>
-         *
-         * @see DefaultBearerProvider
-         */
-        private boolean forwardBearer = false;
     }
 
     public Optional<OAuth2LogoutProperties> getLogoutProperties(String clientRegistrationId) {