+++ spring-addons-starter-oidc/src/main/java/com/c4_soft/springaddons/security/oidc/starter/properties/condition/HasRestConfiguration.java
@@ -1,8 +1,0 @@
-package com.c4_soft.springaddons.security.oidc.starter.properties.condition;
-
-public class HasRestConfiguration extends HasPropertyPrefixCondition {
-
-    public HasRestConfiguration() {
-        super("com.c4-soft.springaddons.oidc.client.rest");
-    }
-}